// apps/api/src/common/roles.ts
export enum Role {
  SuperAdmin = "SUPER_ADMIN",
  CenterAdmin = "CENTER_ADMIN",
  Operator = "OPERATOR"
}

export enum VerificationStatus {
  Verified = "VERIFIED",
  NotVerified = "NOT_VERIFIED",
  Review = "REVIEW"
}

export enum DeviceStatus {
  Online = "ONLINE",
  Offline = "OFFLINE",
  Maintenance = "MAINTENANCE"
}